requirements.txt
flask
flask-restful
flask-restx
pymongo
torch
chromadb
langchain
langchain-community
langchain_chroma
langchain-huggingface
sentence-transformers
scikit-learn
numpy
python-dotenv